fixed issue resulting in low code cov results
