+++ reductstore/src/storage/entry.rs
@@ -41,7 +41,9 @@ use tokio::sync::OwnedSemaphorePermit;
 use tokio::task::JoinHandle;
 
 pub(crate) use builder::EntryBuilder;
-pub(crate) use pattern::{entry_matches_pattern, matches_patterns};
+pub(crate) use pattern::entry_matches_pattern;
+#[cfg(feature = "zenoh-api")]
+pub(crate) use pattern::matches_patterns;
 
 struct QueryHandle {
     rx: Arc<AsyncRwLock<QueryRx>>,
